+++ lib/go-tc/log.go
@@ -25,13 +25,6 @@ type LogsResponse struct {
 	Alerts
 }
 
-type UserLogResponse struct {
-	Response []Log `json:"response"`
-	Summary  struct {
-		Count uint64 `json:"count"`
-	} `json:"summary"`
-}
-
 // Log contains a change that has been made to the Traffic Control system.
 type Log struct {
 	ID          *int    `json:"id"`